Created custom hook useStarringEpisodes and fixed bugs from PR
